requirements.txt
doxieapi==0.0.2
click==6.2
Django==1.9.1
psycopg2==2.6.1
PyYAML==3.11
